requirements.txt
flask
python-dotenv
requests
sqlite3